==> 2.finetuning.ipynb <==
import os
from tqdm import tqdm 
from rdkit.Chem import QED
from torch.utils.data import Dataset, DataLoader
import numpy as np
import torch
import pandas as pd
import time
from IPython.display import clear_output
import itertools
import importlib
import random
from utils import *

with open("./data/zinc15_selfies_tokens.txt", "r") as file:
    token_in_dataset = [line.strip() for line in file]

print("Loaded tokens:", token_in_dataset)

word2index = {"<pad>": 0, "<unk>": 1, "<sos>": 2, "<eos>": 3}
index2word = {0: "<pad>", 1: "<unk>", 2: "<sos>", 3: "<eos>"}

start_index = max(index2word.keys()) + 1

for i, token in enumerate(token_in_dataset, start=start_index):
    word2index[token] = i
    index2word[i] = token

print("word2index:", word2index)
print("index2word:", index2word)

data_path = './data/zinc15_sample.csv'
df = pd.read_csv(data_path)
smiles = df['smiles'].to_numpy()
selfies = df['selfies'].to_numpy()

prop_num = 3
prop_data = [df['bcl2'].to_numpy(),
             df['bclxl'].to_numpy(),
             df['bclw'].to_numpy()]
prop_name = ['bcl2', 'bclxl', 'bclw']

GPU_NUM = 0
device = torch.device(f'cuda:{GPU_NUM}' if torch.cuda.is_available() else 'cpu')
torch.cuda.set_device(device)

dataset = selfiesDataset(selfies, prop_data, word2index, device, num_samples=None)
data_loader = DataLoader(dataset,
                     batch_size=128,
                     shuffle=True,
                     collate_fn=lambda x: collate_fn(x, word2index, dataset.pattern, device))
data_len = len(dataset)
print(data_len)

from tqdm.notebook import tqdm
from tqdm.auto import trange

num_epoch = 100

model_name = 'finetuned_vae_zinc15'
save_path = f'./model/{model_name}/'
if not os.path.exists(save_path):
    os.makedirs(save_path)

from VAE import *

embed_dim = 256                   #Embedding Vector Dim 
hidden_dim = 512                   #Latent Vector Dim
latent_dim = 256
en_n_l = 3                         #Encoder GRU Number of Layers 
de_n_l = 2                         #Decoder GRU Number of Layers
base_batch_size = 128                   # Batch Size of training data
learning_rate = 1e-4

from VAE import *
model = VAE(voca_dim=len(word2index),
            embed_dim=embed_dim,
            hidden_dim=hidden_dim,
            latent_dim=latent_dim,
            en_num_layers=en_n_l,
            de_num_layers=de_n_l,
            prop_num = prop_num,
            run_predictor = True,
            value_range = None).to(device)

pt_path = f'./model/pretrained_vae_zinc15.pt'
state_dict = torch.load(pt_path)['model_state_dict']
filtered_state_dict = {k: v for k, v in state_dict.items() if not (k.startswith('predictors') or k.startswith('latent_mask'))}
model.load_state_dict(filtered_state_dict, strict=False)


NLL = nn.NLLLoss(reduction='none',
                 ignore_index=0)  # Reconstruct Loss 
GaussianNLL = torch.nn.GaussianNLLLoss(reduction="sum")
optimizer = optim.Adam(model.parameters(), lr=learning_rate)

epoch_log = []
iterate_log = []
loss_log = []
rec_loss_log = []
kl_loss_log = []
prop_loss_log = []
ori_kl_loss_log = []
ori_prop_loss_log = []

total_loss = 0
is_print = True

kl_w_rate = 0.1
prop_weight = [10, 10, 10] # Model A is Same Weight  
sos_tensor = torch.tensor([word2index['<sos>']]).to(device)
start_time = time.time()
non_zero = 1e-8  # add this when calculate log() e.g. log(std + non_zero)
i = -1    

start_epoch = 0
num_epoch = 100

for epoch in trange(start_epoch, num_epoch, desc="Epcohs", leave=False):    
    for batch in tqdm(data_loader, desc=f"Dataset({data_len})/batch({base_batch_size})", leave=False):
        model.train()
        i += 1
        one_time = time.time()
        sorted_source, sorted_target, sorted_lengths, max_len, sorted_props, sorted_origin_indexs = batch
        batch_size = len(sorted_source)
        src = sorted_source.to(device)# ['<sos>', O', '=', 'C']
        trg = sorted_target.to(device)# ['O', '=', 'C', '<eos>']
    
    
        tf_ratio = 0.0
        x, y, z, sample_z, mask = model.forward(src, trg, sorted_lengths, max_len, tf_ratio = tf_ratio)
        x_ = x.view(-1, x.shape[2]).contiguous()
        x_label = sorted_target.view(-1).contiguous().to(device) 
        
        # Reconstruct Loss
        rec_loss = NLL(x_, x_label)  # rec_loss = F.binary_cross_entropy(x,  )
        rec_loss = rec_loss.view(x.shape[0], x.shape[1])
        rec_loss = torch.sum(rec_loss, dim=-1) # [batch, token_num ] -> [batch]
        rec_loss = torch.sum(rec_loss) / batch_size
    
        # KLD Loss 
        
        mu, std = z
        kl_loss = torch.mean(mu ** 2 + std ** 2 - 2 * torch.log(std + non_zero) - 1, dim=1)
        ori_kl_loss = torch.mean(kl_loss).item()
        kl_loss = kl_w_rate * torch.mean(kl_loss)
    
        prop_loss = []
        props_to_iterate = sorted_props
        for idx, y_label in enumerate(props_to_iterate):
            y_label_float = torch.tensor(y_label, dtype=torch.float).unsqueeze(1).to(device)
        
            predicted_mean = y[idx]  
            target = y_label_float  
            variance = torch.ones_like(predicted_mean) 
        
            # Gaussian NLL Loss
            y_loss = GaussianNLL(predicted_mean, target, variance) / batch_size
            prop_loss.append(y_loss) 
            
        weight_prop_loss = [(prop_loss[k] * prop_weight[k]) / (prop_num) for k in range(prop_num)]
        sum_prop_loss = sum(weight_prop_loss)
        loss = rec_loss + kl_loss + sum_prop_loss 
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
        total_loss += loss.item()
    
        # Log 
        epoch_log.append(epoch)
        iterate_log.append(i)
        loss_log.append(loss.item())
        rec_loss_log.append(rec_loss.item())
        kl_loss_log.append(kl_loss.item())
        ori_kl_loss_log.append(ori_kl_loss)
        prop_loss_log.append(sum_prop_loss.item())
        ori_prop_loss_log.append(sum(prop_loss).item())
        if i % 100 == 0:
            if is_print:
                with print_out:
                    true_false_counts = [
                        (torch.sum(mask[0, i]).item(), (mask[0, i].numel() - torch.sum(mask[0, i])).item())
                        for i in range(mask.size(1))
                    ]
                    mask_true = sum(true_count for true_count, _ in true_false_counts)
                    mask_false = sum(false_count for _, false_count in true_false_counts)
                    clear_output(wait=True)
                    print(f'=========================epoch : {epoch}==========================')
                    print(f'# Model Name : {model_name}')
                    print("Iteration ", i, ", Total loss ", loss.item(), "\nKL loss ", kl_loss.item(),
                          ", Rec loss ", rec_loss.item(), sep="")
                    print("Origin_KL_loss : ", ori_kl_loss)
                    print("Origin_prop_loss : ", sum(prop_loss).item())
                    print(f'Mask True : " {mask_true}')
                    print(f'Mask False : " {mask_false}')
                    print(f'sum(y Loss) : {sum_prop_loss.item()}')
                    for p_i in range(prop_num):
                        print(f'== = Property : {prop_name[p_i]} = ==')
                        p_true, p_false = true_false_counts[p_i]
                        print(f'Mask True Num : {p_true}')
                        print(f'Mask False Num : {p_false}')
                        print(f'y[{p_i}] Loss : {prop_loss[p_i].item()}')
                        print(f'Fixed_dynamic_weight: {prop_weight[p_i]}')
                        print(f'input_{prop_name[p_i]}[{idx}] : {sorted_props[p_i][idx]}')
                        print(f'pred_{prop_name[p_i]}[{idx}] : {y[p_i][idx].item()}')
                        print("-----------------------------")
    
                    print("Time: ", convert_time(time.time() - one_time))
                    print("Teacher Forcing Ratio : ", tf_ratio)
                    print("KLD weight : ", kl_w_rate)
                    
                    idx = print_token2sf(sorted_target, x, word2index, index2word, batch_size, num=1)
                    idx = idx[0].item()
                    print(" idx : ",idx)
                    print("-----")
                    _, tmp_a = torch.max(x[idx], dim=-1)
                    tmp_a = tmp_a.reshape(-1).cpu().detach().numpy()
                    in_sentence = [index2word[i] for i in tmp_a]
                    in_smi = ''.join(in_sentence)
                    print(in_smi)
                    print("-----")
                    print("===============================================================")
                one_time = time.time()
            torch.save({'model_state_dict': model.state_dict(),
                        'optimizer_state_dict': optimizer.state_dict(),
                        }, f'{save_path}{model_name}_main.pt')
    
            df = pd.DataFrame({'epoch': epoch_log,
                               'i': iterate_log,
                               'loss': loss_log,
                               'rec_loss': rec_loss_log,
                               'kl_loss': kl_loss_log,
                               'prop_loss': prop_loss_log,
                               'ori_kl_loss' : ori_kl_loss_log,
                               'ori_prop_loss' : ori_prop_loss_log})
            df.to_csv(f'{save_path}Log_{model_name}.csv',
                      index=False)
            if (epoch + 1) % 5 == 0:
                torch.save({'model_state_dict': model.state_dict(),
                'optimizer_state_dict': optimizer.state_dict(),
                }, f'{save_path}{model_name}_epoch_{epoch}.pt')
                # Decrease tf_ratio for the next epoch 
mean_iteration_loss = total_loss / i
print("===============================================================")
print("Completed Epoch", epoch, ", Total loss Mean: ", mean_iteration_loss, ", Time: ",
      convert_time(time.time() - start_time))
print("===============================================================")
