==> 3.optimize.ipynb <==
import os
from tqdm import tqdm 
from rdkit.Chem import QED
from torch.utils.data import Dataset, DataLoader
import numpy as np
import torch
import pandas as pd
import time
from IPython.display import clear_output
import itertools
import importlib
import random
from utils import *

with open("./data/zinc15_selfies_tokens.txt", "r") as file:
    token_in_dataset = [line.strip() for line in file]

print("Loaded tokens:", token_in_dataset)
word2index = {"<pad>": 0, "<unk>": 1, "<sos>": 2, "<eos>": 3}
index2word = {0: "<pad>", 1: "<unk>", 2: "<sos>", 3: "<eos>"}

start_index = max(index2word.keys()) + 1

for i, token in enumerate(token_in_dataset, start=start_index):
    word2index[token] = i
    index2word[i] = token
    
print("word2index:", word2index)
print("index2word:", index2word)

data_path = './data/zinc15_sample_test.csv'
df = pd.read_csv(data_path)
smiles = df['smiles'].to_numpy()
selfies = df['selfies'].to_numpy()

prop_num = 3
prop_data = [df['bcl2'].to_numpy(),
             df['bclxl'].to_numpy(),
             df['bclw'].to_numpy()]

prop_name = ['bcl2', 'bclxl', 'bclw']

embed_dim = 256                   #Embedding Vector Dim 
hidden_dim = 512                   #Latent Vector Dim
latent_dim = 256
en_n_l = 3                         #Encoder GRU Number of Layers 
de_n_l = 2                         #Decoder GRU Number of Layers
base_batch_size = 128                   # Batch Size of training data
learning_rate = 1e-4

GPU_NUM = 0
device = torch.device(f'cuda:{GPU_NUM}' if torch.cuda.is_available() else 'cpu')
torch.cuda.set_device(device)


from VAE import *
model = VAE(voca_dim=len(word2index),
            embed_dim=embed_dim,
            hidden_dim=hidden_dim,
            latent_dim=latent_dim,
            en_num_layers=en_n_l,
            de_num_layers=de_n_l,
            prop_num = prop_num,
            run_predictor = True,
            value_range = None).to(device)
pt_path = f'./model/finetuned_Model_A.pt'
state_dict = torch.load(pt_path)['model_state_dict']
model.load_state_dict(state_dict)
model.eval()

target_score_dict = {'qed' : (0.8, 'up', {'min' : 0, 'max' : 1}),
                     'SAs' : (3.0, 'down', {'min' : 1, 'max' : 10}), 
                     'bcl2' : (10.34, 'up', {'min' : 2.15, 'max' : 10.34}),
                     'bclxl' : (9.85, 'up', {'min' : 1.91, 'max' : 9.85}),
                     'bclw' : (8.68, 'up', {'min' : 2.23, 'max' : 8.68})}
target_list = [ target_score_dict[str][0] for str in prop_name] 
target_up = [ True if target_score_dict[str][1] == 'up' else False for str in prop_name]  #Target up
target_tensor = torch.tensor(target_list).to(device)
target_max = [ target_score_dict[str][2]['max'] for str in prop_name ]
target_min = [ target_score_dict[str][2]['min'] for str in prop_name ]



final_df = df.sample(n=100, random_state=2025)
smiles = final_df['smiles'].to_numpy()
selfies = final_df['selfies'].to_numpy()

prop_data = []
for i in range(prop_num):
    prop_data.append(final_df[prop_name[i]].to_numpy())

dataset = selfiesDataset(selfies, prop_data, word2index, device, num_samples=None)
data_loader = DataLoader(dataset,
                         batch_size=1,
                         shuffle=False,
                         collate_fn=lambda x: collate_fn(x, word2index, dataset.pattern, device))
print(final_df[['smiles']+ prop_name][0:3])

#1. define odds weight  by importance sampling 
tau = 0.4

#Get MPL prop
_, odds_full = model.mask_inference(thold=0.0)   # odds_full: (prop_num, latent_dim, 2)
odds = odds_full[:, :, 1]   

#odds_ratio 
odds_scaled = odds ** (1.0 / tau)

weights = odds_scaled / (odds_scaled.mean(dim=1, keepdim=True) + 1e-8)
weights += 1.0   
weight_vectors = [weights[i].clone() for i in range(weights.size(0))]


#Log 저장 경로
model_name = f"Model_A_Optim"
save_optim_path = f'./model/{model_name}/'

parent_num = 10
stop_condi = 3000

t_weight = 15
prop_weight = [1, 1, 1]

MSE = torch.nn.MSELoss(reduction="sum")
GaussianNLL = torch.nn.GaussianNLLLoss(reduction="sum")

optim_step = 50
optim_iter = 20  

y_list = []
for p_i, batch in enumerate(data_loader):
    # Define Model
    
    import VAE as vae
    importlib.reload(vae)
    from VAE import *
    model = VAE(voca_dim=len(word2index),
                embed_dim=embed_dim,
                hidden_dim=hidden_dim,
                latent_dim=latent_dim,
                en_num_layers=en_n_l,
                de_num_layers=de_n_l,
                prop_num = prop_num,
                run_predictor = True,
                value_range = None).to(device)
    pt_path = f'./model/finetuned_Model_A.pt'
    state_dict = torch.load(pt_path)['model_state_dict']
    model.load_state_dict(state_dict)
    model.eval()
    for param in model.parameters():
        param.requires_grad = False

    #\eta parameter
    z_param = nn.Parameter(torch.randn(1, latent_dim), requires_grad=True)

    sorted_source, sorted_target, sorted_lengths, max_len, sorted_props, sorted_origin_indexs = batch
    batch_size = len(sorted_source)
    src = sorted_source.to(device)
    trg = sorted_target.to(device)
    x, y, _, sample_z, _ = model.inference(src,
                                           sorted_lengths,
                                           max_len)

    z_param.data = sample_z.clone().data  
    z_param.requires_grad = True

    optimizer = optim.Adam([z_param], lr=0.01)
    sos_tensor = torch.tensor([word2index['<sos>']]).to(device)
    smi_src, sf = tensor2smiles_v2(x, word2index, index2word)
    smi = smi_src[0]

    _, y = model.decode_z(z_param, max_len = 154, sos_tensor=sos_tensor, return_y = True) 
    y_values = [tensor_item.cpu().detach().numpy().flatten()[0] for tensor_item in y]
    y_list.append(y_values)
    
    for gen_try in range(optim_step * optim_iter):
        individual_losses = []
        ori_t_losses = []
        pred_list = []
        
        x = model.decode_z(z_param, max_len = 154, sos_tensor=sos_tensor) 
        smi_src, sf = tensor2smiles_sampling(x, word2index, index2word)  
        smi = normalize_SMILES(smi_src[0])
        optimizer.zero_grad() 
        total_grad_z = torch.zeros_like(z_param) 

        for idx, (predictor, target, is_up, p_max, p_min) in enumerate(zip(model.predictors, target_tensor, target_up, target_max, target_min)):
            pred_y = predictor(z_param)
            pred_list.append(pred_y)
            t_loss = MSE(pred_y.squeeze(0).view(target.shape), target)
            individual_losses.append(prop_weight[idx] * t_loss)
            ori_t_losses.append(t_loss)
    
            weight_vector = weight_vectors[idx]
            weight_vector = weight_vector.to(dtype=z_param.dtype, device=z_param.device)
    
            grad_z = torch.autograd.grad(prop_weight[idx] * t_loss, z_param, retain_graph=True)[0]
    
            weighted_grad_z = grad_z * weight_vector.unsqueeze(0)
    
            total_grad_z += weighted_grad_z

        target_loss = torch.mean(torch.stack(individual_losses))
        loss = (t_weight * target_loss)
        z_param.grad = total_grad_z
        optimizer.step()        
        #loss.backward()

        if (gen_try % optim_iter == 0):         
            _, y = model.decode_z(z_param, max_len = 154, sos_tensor=sos_tensor, return_y = True) 
            y_values = [tensor_item.cpu().detach().numpy().flatten()[0] for tensor_item in y]
            y_list.append(y_values)

            
            clear_output(wait=True)
            print(f'= =step : {int(gen_try / optim_iter)}, seed {p_i} Optimize Results === = = =')
            print("Current Step Z mean:", z_param.data.mean().item())
            print("#" * 20)
            print("smi : ", smi)
            print("Target Loss : ", t_weight * target_loss.item())

            for i in range(prop_num):
                print('#'* 20 )
                print(f'# Property Info {i}  - {prop_name[i]} -  # ')
                print(f'Property Range | {target_min[i]}(min) ~ {target_max[i]}(max) |')
                print(f'Target : {target_list[i]}')
                print(f'Predict : {pred_list[i].item()}')

            print(f'= = = === = = = === = = = === = = = === = = = === = = =')
    mask_y_list = list(map(list, zip(*y_list)))

print("End of Optimization")
